server/src/requirements.txt
python-dotenv
flask
flask_sqlalchemy
flask_cors
psycopg2
pytz
pyjwt
cryptography
python-json-logger